🛠️ refactor: (FE) - feedback 관련 컴포넌트 구조 및 이름 변경
