+++ frontend/src/components/FeedbackItem/FeedbackItem.style.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 
-export const feedbackBoxStyle = (theme) => css`
+export const feedbackBoxStyle = css`
 	display: flex;
 	gap: 10px;
 	position: relative;
@@ -14,16 +14,6 @@ export const fbTextAreaStyle = (theme) => css`
 	background-color: ${theme.colors.secondary};
 	padding: 15px 30px 15px 15px;
 	box-sizing: border-box;
-`;
-
-export const fbBtnContainer = css`
-	position: absolute;
-	right: 10px;
-	height: 100%;
-	padding: 5px 0px;
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
 `;
 
 export const fbStartTimeStyle = (theme) => css`
