+++ frontend/src/components/FeedbackItem/FeedbackItem.tsx
@@ -5,16 +5,8 @@ import useCrudFeedback from '@hooks/useCrudFeedback';
 import { isFbClickedState, isFbSyncState } from '@store/feedback.store';
 import { currentVideoTimeState } from '@store/currentVideoTime.store';
 
-import { ReactComponent as DeleteIcon } from '@assets/icon/delete.svg';
-import { ReactComponent as EditIcon } from '@assets/icon/edit.svg';
-import { ReactComponent as CheckIcon } from '@assets/icon/check.svg';
-import {
-	feedbackBoxStyle,
-	fbTextAreaStyle,
-	fbBtnContainer,
-	fbStartTimeStyle,
-} from './EditableFeedbackBox.style';
-import { iconSmStyle } from '@styles/commonStyle';
+import { feedbackBoxStyle, fbTextAreaStyle, fbStartTimeStyle } from './FeedbackItem.style';
+
 import { FeedbackItemType } from '@customType/feedback';
 
 interface PropsType {
@@ -23,13 +15,12 @@ interface PropsType {
 	feedbackRef: React.MutableRefObject<any[]>;
 	index: number;
 }
-const EditableFeedbackBox = ({ feedback, feedbackRef, index }: PropsType) => {
+const FeedbackItem = ({ feedback, feedbackRef, index }: PropsType) => {
 	const textareaRef = useRef<HTMLTextAreaElement>(null);
 	const isFbSync = useRecoilValue(isFbSyncState);
 	const setIsFbClicked = useSetRecoilState(isFbClickedState);
 	const setCurrentVideoTime = useSetRecoilState(currentVideoTimeState);
-	const { handleStartEditFeedback, handleEndEditFeedback, handleFbChange, handleDeleteFeedback } =
-		useCrudFeedback(feedback.id);
+	const { handleFbChange } = useCrudFeedback(feedback.id);
 
 	useEffect(() => {
 		feedbackRef.current[index].style.height = textareaRef.current.scrollHeight + 'px';
@@ -56,22 +47,8 @@ const EditableFeedbackBox = ({ feedback, feedbackRef, index }: PropsType) => {
 				onClick={handleClickFeedback}
 				css={fbTextAreaStyle}
 			/>
-			<div css={fbBtnContainer}>
-				{readOnly ? (
-					<button onClick={handleStartEditFeedback}>
-						<EditIcon {...iconSmStyle} fill="black" />
-					</button>
-				) : (
-					<button onClick={handleEndEditFeedback}>
-						<CheckIcon {...iconSmStyle} fill="black" />
-					</button>
-				)}
-				<button onClick={handleDeleteFeedback}>
-					<DeleteIcon {...iconSmStyle} fill="black" />
-				</button>
-			</div>
 		</div>
 	);
 };
 
-export default React.memo(EditableFeedbackBox);
+export default React.memo(FeedbackItem);
