+++ frontend/src/components/FeedbackList/FeedbackList.style.ts
@@ -1,0 +1,9 @@
+import { css } from '@emotion/react';
+
+export const feedbackListStyle = css`
+	overflow: scroll;
+	display: flex;
+	flex-direction: column;
+	height: 70%;
+	gap: 10px;
+`;
